+++ src/miceFast_quantmodels.cpp
@@ -3,32 +3,42 @@
 
 #include <RcppArmadillo.h>
 #include "miceFast.h"
+#include <stdlib.h>     /* rand */
+#include <algorithm>
+using namespace std;
+using namespace Rcpp;
+using namespace arma;
+
 
 #define UNUSED(expr) (void)(expr)
 
 //weighted linear regression
-arma::colvec fastLm_weighted(arma::colvec &y, arma::mat &X,arma::colvec &w, arma::mat &X1,int times) {
+arma::colvec fastLm_weighted(arma::colvec &y, arma::mat &X,arma::colvec &w, arma::mat &X1,int k, double ridge) {
 
-  UNUSED(times);
+  UNUSED(k);
 
-  int n = X.n_rows, k = X.n_cols;
+  int n = X.n_rows;
+  int C = X.n_cols;
 
   arma::colvec wq=sqrt(w);
   arma::colvec y2 = wq%y;
-  arma::mat X2(n,k);
+  arma::mat X2(n,C);
 
-  for(int h=0;h<k;h++){
+  for(int h=0;h<C;h++){
     X2.col(h)=wq%X.col(h);
   }
 
-  arma::colvec coef = arma::solve(X2.t()*X2, X2.t()*y2);
+  arma::mat XX = X2.t()*X2 ;
+  XX.diag() += ridge;
+
+  arma::colvec coef = arma::inv(XX)*X2.t() * y2;
 
   return X1*coef;
 
 }
 
 //weighted linear regression - noise
-arma::colvec fastLm_weighted_noise(arma::colvec &y, arma::mat &X,arma::colvec &w,arma::mat &X1,int times) {
+arma::colvec fastLm_weighted_noise(arma::colvec &y, arma::mat &X,arma::colvec &w,arma::mat &X1,int k, double ridge) {
 
   int N = X.n_rows; int C = X.n_cols; int N_NA = X1.n_rows;
 
@@ -40,7 +50,10 @@ arma::colvec fastLm_weighted_noise(arma::colvec &y, arma::mat &X,arma::colvec &w
     X2.col(h)=wq%X.col(h);
   }
 
-  arma::colvec coef = arma::solve(X2.t()*X2, X2.t()*y2);
+  arma::mat XX = X2.t()*X2 ;
+  XX.diag() += ridge;
+
+  arma::colvec coef = arma::inv(XX)*X2.t() * y2;
 
   arma::colvec res = y - X*coef;
 
@@ -48,7 +61,7 @@ arma::colvec fastLm_weighted_noise(arma::colvec &y, arma::mat &X,arma::colvec &w
 
   arma::colvec pred_sum(N_NA,arma::fill::zeros);
 
-  for(int i=0;i<times;i++){
+  for(int i=0;i<k;i++){
 
     arma::vec noise(N_NA);
     noise.randn();
@@ -56,15 +69,15 @@ arma::colvec fastLm_weighted_noise(arma::colvec &y, arma::mat &X,arma::colvec &w
     pred_sum = pred_sum + (X1 * coef + noise * sigma);
   }
 
-  return pred_sum/(double) times;
+  return pred_sum/(double) k;
 
 }
 
 //weighted linear regression - bayes
 
-arma::colvec fastLm_weighted_bayes(arma::colvec &y, arma::mat &X,arma::colvec &w,arma::mat &X1,int times) {
+arma::colvec fastLm_weighted_bayes(arma::colvec &y, arma::mat &X,arma::colvec &w,arma::mat &X1,int k, double ridge) {
 
-  int N = X.n_rows; int C = X.n_cols; int N_NA = X1.n_rows;
+  int N = X.n_rows; int C = X.n_cols; int N_NA = X1.n_rows; int C_NA = X1.n_cols;
 
   arma::colvec wq=sqrt(w);
   arma::colvec y2 = wq%y;
@@ -74,19 +87,24 @@ arma::colvec fastLm_weighted_bayes(arma::colvec &y, arma::mat &X,arma::colvec &w
     X2.col(h)=wq%X.col(h);
   }
 
-  arma::colvec coef = arma::solve(X2.t()*X2, X2.t()*y2);
+  arma::mat XX = X2.t()*X2 ;
+  XX.diag() += ridge;
+
+  arma::mat XX_inv = arma::inv(XX) ;
+
+  arma::mat XX_inv_chol = arma::chol(XX_inv) ;
+
+  arma::colvec coef = XX_inv*X2.t() * y2;
 
   arma::colvec res = y - X*coef;
 
   double df = N-C ;
 
-  //arma::mat XX_inv = arma::inv(arma::trans(X)*X) ;
-
   double res2 = arma::as_scalar(arma::trans(res)*res);
 
   arma::colvec pred_sum(N_NA,arma::fill::zeros);
 
-  for(int i=0;i<times;i++){
+  for(int i=0;i<k;i++){
 
     double chi2 = Rcpp::as<double>(Rcpp::rchisq(1, df));
 
@@ -94,33 +112,43 @@ arma::colvec fastLm_weighted_bayes(arma::colvec &y, arma::mat &X,arma::colvec &w
 
     arma::vec noise2(N_NA);
     noise2.randn();
+    arma::colvec noise2b(C_NA);
+    noise2b.randn();
 
-    pred_sum = pred_sum +( X1 * coef + noise2 * sigma_b);
+    arma::colvec coef2 =  coef + arma::trans(XX_inv_chol) * noise2b * sigma_b;
+    coef2.replace(datum::nan, 0);
+
+    pred_sum = pred_sum + (X1 * coef2 + noise2 * sigma_b);
   }
 
-  return pred_sum/(double) times;
+  return pred_sum/(double) k;
 
 }
 
 //linear regression - bayes
 
-arma::colvec fastLm_bayes(arma::colvec &y, arma::mat &X, arma::mat &X1,int times) {
+arma::colvec fastLm_bayes(arma::colvec &y, arma::mat &X, arma::mat &X1,int k, double ridge) {
 
-  int N = X.n_rows; int C = X.n_cols; int N_NA = X1.n_rows;
+  int N = X.n_rows; int C = X.n_cols; int N_NA = X1.n_rows; int C_NA = X1.n_cols;
 
-  arma::colvec coef =  arma::solve(X.t()*X, X.t()*y);
+  arma::mat XX = X.t()*X ;
+  XX.diag() += ridge;
+
+  arma::mat XX_inv = arma::inv(XX) ;
+
+  arma::mat XX_inv_chol = arma::chol(XX_inv);
+
+  arma::colvec coef = XX_inv*X.t() * y;
 
   arma::colvec res = y - X*coef;
 
   double df = N-C;
 
-  //arma::mat XX_inv = arma::inv(arma::trans(X)*X) ;
-
   double res2 = arma::as_scalar(arma::trans(res)*res);
 
   arma::colvec pred_sum(N_NA,arma::fill::zeros);
 
-  for(int i=0;i<times;i++){
+  for(int i=0;i<k;i++){
 
     double chi2 = Rcpp::as<double>(Rcpp::rchisq(1, df));
 
@@ -128,20 +156,28 @@ arma::colvec fastLm_bayes(arma::colvec &y, arma::mat &X, arma::mat &X1,int times
 
     arma::vec noise2(N_NA);
     noise2.randn();
+    arma::colvec noise2b(C_NA);
+    noise2b.randn();
 
-    pred_sum = pred_sum + (X1 * coef + noise2 * sigma_b);
+    arma::colvec coef2 =  coef + arma::trans(XX_inv_chol) * noise2b * sigma_b;
+    coef2.replace(datum::nan, 0);
+
+    pred_sum = pred_sum + (X1 * coef2 + noise2 * sigma_b);
   }
 
-  return pred_sum/(double) times;
+  return pred_sum/(double) k;
 
 }
 
 //Linear regression with noise
-arma::colvec fastLm_noise(arma::colvec &y,arma::mat &X, arma::mat &X1,int times) {
+arma::colvec fastLm_noise(arma::colvec &y,arma::mat &X, arma::mat &X1,int k, double ridge) {
 
   int N = X.n_rows; int C = X.n_cols; int N_NA = X1.n_rows;
 
-  arma::colvec coef =  arma::solve(X.t()*X, X.t()*y);
+  arma::mat XX = X.t()*X ;
+  XX.diag() += ridge;
+
+  arma::colvec coef = arma::inv(XX)*X.t() * y;
 
   arma::colvec res = y - X*coef;
 
@@ -149,7 +185,7 @@ arma::colvec fastLm_noise(arma::colvec &y,arma::mat &X, arma::mat &X1,int times)
 
   arma::colvec pred_sum(N_NA,arma::fill::zeros);
 
-  for(int i=0;i<times;i++){
+  for(int i=0;i<k;i++){
 
     arma::vec noise(N_NA);
     noise.randn();
@@ -157,17 +193,20 @@ arma::colvec fastLm_noise(arma::colvec &y,arma::mat &X, arma::mat &X1,int times)
     pred_sum = pred_sum + (X1 * coef + noise * sigma);
   }
 
-  return pred_sum/(double) times;
+  return pred_sum/(double) k;
 
 }
 
 //Simple linear regression
 
-arma::colvec fastLm_pred(arma::colvec &y, arma::mat &X, arma::mat &X1,int times) {
+arma::colvec fastLm_pred(arma::colvec &y, arma::mat &X, arma::mat &X1,int k, double ridge) {
 
-  UNUSED(times);
+  UNUSED(k);
 
-  arma::colvec coef = arma::solve(X.t()*X, X.t()*y);
+  arma::mat XX = X.t()*X ;
+  XX.diag() += ridge;
+
+  arma::colvec coef = arma::inv(XX)*X.t() * y;
 
   return X1*coef;
 
@@ -175,9 +214,9 @@ arma::colvec fastLm_pred(arma::colvec &y, arma::mat &X, arma::mat &X1,int times)
 
 //LDA prediction model
 
-arma::colvec fastLda( arma::colvec &y,  arma::mat &X, arma::mat &X1, int times) {
+arma::colvec fastLda( arma::colvec &y,  arma::mat &X, arma::mat &X1, int k, double ridge) {
 
-  UNUSED(times);
+  UNUSED(k);
 
   arma::uvec vars = arma::find(arma::var(X)>0);
 
@@ -225,8 +264,8 @@ arma::colvec fastLda( arma::colvec &y,  arma::mat &X, arma::mat &X1, int times) 
   arma::mat X0 = sqrt(fac) * (X_vol - group_means_mat) * scaling;
 
   X_vol.clear();
-
-  arma::mat input = X0.t()*X0;
+  arma::mat input = X0.t()*X0 ;
+  input.diag() += ridge;
   X0.clear();
   arma::mat U;
   arma::mat V;
@@ -306,4 +345,266 @@ arma::colvec fastLda( arma::colvec &y,  arma::mat &X, arma::mat &X1, int times) 
 
 //Ridge prediction model
 
+//PMM
 
+int findCrossOver(arma::colvec arr, double low, double high, double x)
+{
+  if (arr[high] <= x)
+    return high;
+  if (arr[low] > x)
+    return low;
+
+
+  int mid = (low + high)/2;
+
+  if (arr[mid] <= x && arr[mid+1] > x)
+    return mid;
+
+  if(arr[mid] < x)
+    return findCrossOver(arr, mid+1, high, x);
+
+  return findCrossOver(arr, low, mid - 1, x);
+}
+
+
+arma::colvec Kclosestrand(arma::colvec arr, double x, int k)
+{
+  int n = arr.n_rows;
+
+  int l = findCrossOver(arr, 0, n-1, x);
+  int r = l;
+  int count = 0;
+  arma::colvec resus(k);
+
+  if (arr[l] == x) l--;
+
+  while (l >= 0 && r < n && count < k)
+  {
+    if (x - arr[l] < arr[r] - x)
+      resus[count] = arr[l--];
+    else
+      resus[count] = arr[r++];
+    count++;
+  }
+
+  while (count < k && l >= 0)
+    resus[count] = arr[l--], count++;
+
+  while (count < k && r < n)
+    resus[count] = arr[r++], count++;
+
+  return resus;
+
+}
+
+
+
+//' Finding in random manner one of the k closets points in a certain vector for each value in a second vector
+//'
+//' @description this function using pre-sorting of a y and the binary search the one of the k closest value for each miss is returned.
+//'
+//' @param y numeric vector values to be look up
+//' @param miss numeric vector a values to be look for
+//' @param k integer a number of values which should be taken into account during sampling one of the k closest point
+//'
+//' @return a numeric vector
+//'
+//' @name neibo
+//'
+//' @export
+// [[Rcpp::export]]
+arma::colvec neibo(arma::colvec y, arma::colvec miss, int k) {
+
+  int n_y = y.n_rows;
+
+  k = (k <= n_y) ? k : n_y;
+  k = (k >= 1) ? k : 1;
+
+  arma::colvec y_new = arma::sort(y);
+
+  int n_miss = miss.n_rows;
+
+  arma::colvec resus(n_miss);
+  arma::colvec resus_temp(k);
+
+  arma::vec which = floor(runif(n_miss, 0, k));
+
+
+    int subs;
+
+  for(int i=0; i<n_miss ;i++){
+    double mm = miss[i];
+    resus_temp = Kclosestrand(y_new,mm,k);
+    subs = (int) which[i];
+    resus[i] = resus_temp[subs];
+  }
+
+  return resus ;
+
+}
+
+
+
+arma::uvec Kclosestrand_index(arma::colvec arr, double x, int k)
+{
+  int n = arr.size();
+
+  int l = findCrossOver(arr, 0, n-1, x);
+  int r = l;
+  int count = 0;
+  arma::uvec resus(k);
+
+  if (arr[l] == x) l--;
+
+  while (l >= 0 && r < n && count < k)
+  {
+    if (x - arr[l] < arr[r] - x)
+      resus[count] = l--;
+    else
+      resus[count] = r++;
+    count++;
+  }
+
+  while (count < k && l >= 0)
+    resus[count] = l--, count++;
+
+  while (count < k && r < n)
+    resus[count] = r++, count++;
+
+  return resus;
+
+}
+
+
+
+//' Finding in random manner one of the k closets points in a certain vector for each value in a second vector
+//'
+//' @description this function using pre-sorting of a \code{y} and the by the binary search the one of the k closest value for each miss is returned.
+//'
+//' @param y numeric vector values to be look up
+//' @param miss numeric vector a values to be look for
+//' @param k integer a number of values which should be taken into account during sampling one of the k closest point
+//'
+//' @return a numeric vector
+//'
+//' @name neibo_index
+//'
+//' @export
+// [[Rcpp::export]]
+arma::uvec neibo_index(arma::colvec y, arma::colvec miss, int k) {
+  int n_y = y.n_rows;
+
+  k = (k <= n_y) ? k : n_y;
+  k = (k >= 1) ? k : 1;
+
+  arma::colvec y_new = arma::sort(y);
+
+  int n_miss = miss.n_rows;
+
+  arma::uvec resus(n_miss);
+  arma::uvec resus_temp(k);
+
+  arma::vec which = floor(runif(n_miss, 0, k));
+
+
+  int subs;
+
+  for(int i=0; i<n_miss ;i++){
+    double mm = miss[i];
+    resus_temp = Kclosestrand_index(y_new,mm,k);
+    subs = (int) which[i];
+    resus[i] = resus_temp[subs];
+  }
+
+  return resus ;
+
+}
+
+
+// [[Rcpp::export]]
+arma::colvec pmm_weighted_neibo( arma::colvec &y, arma::mat &X,arma::colvec &w,arma::mat &X1,int k, double ridge) {
+
+  int N = X.n_rows; int C = X.n_cols; int C_NA = X1.n_cols;int N_NA = X1.n_rows;
+
+  arma::colvec wq=sqrt(w);
+  arma::colvec y2 = wq%y;
+  arma::mat X2(N,C);
+
+  for(int h=0;h<C;h++){
+    X2.col(h)=wq%X.col(h);
+  }
+
+
+    arma::mat xtx = arma::mat( arma::trans(X2) * X2);
+    for (int ii=0;ii<C;ii++){
+        xtx(ii,ii)=xtx(ii,ii)+ridge;
+    }
+    arma::mat xinv = arma::inv( xtx );
+    arma::vec coef2 = arma::mat( xinv * arma::trans(X2) * y2 );
+    arma::colvec resid = arma::mat( y2 - X2*coef2 );
+    double res2 = arma::as_scalar(resid.t()*resid);
+
+    double df = N - C;
+
+    double chi2 = Rcpp::as<double>(Rcpp::rchisq(1, df));
+
+    double sigma_b = sqrt(res2/chi2);
+
+    arma::vec noise2(N_NA);
+    noise2.randn();
+    arma::colvec noise2b(C_NA);
+    noise2b.randn();
+
+    arma::colvec coef3 =  coef2 + arma::trans(arma::chol(xinv)) * noise2b * sigma_b;
+       coef3.replace(datum::nan, 0);
+
+    arma::colvec ypred_mis = X1 * coef3 + noise2 * sigma_b;
+
+    arma::colvec y_full = arma::sort(y);
+
+    arma::colvec yimp = neibo(y_full,ypred_mis,k);
+
+    return yimp;
+}
+
+arma::colvec pmm_neibo( arma::colvec &y, arma::mat &X,arma::mat &X1,int k, double ridge) {
+
+  int N = X.n_rows; int C = X.n_cols; int C_NA = X1.n_cols;int N_NA = X1.n_rows;
+
+
+    arma::mat xtx = arma::mat( arma::trans(X) * X);
+    xtx.diag() = xtx.diag() + ridge;
+    arma::mat xinv = arma::inv( xtx );
+    arma::vec coef2 =  xinv * arma::trans(X) * y ;
+    arma::colvec resid =  y - X*coef2 ;
+    double res2 = arma::as_scalar(resid.t() *resid);
+
+    double df = N - C;
+
+    double chi2 = Rcpp::as<double>(Rcpp::rchisq(1, df));
+
+    double sigma_b = sqrt(res2/chi2);
+
+    arma::vec noise2(N_NA);
+    noise2.randn();
+    arma::colvec noise2b(C_NA);
+    noise2b.randn();
+
+    arma::colvec coef3 =  coef2 + arma::trans(arma::chol(xinv)) * noise2b * sigma_b;
+    coef3.replace(datum::nan, 0);
+
+    arma::colvec ypred_mis =  X1 * coef3 + noise2 * sigma_b;
+    arma::colvec y_full = arma::sort(y);
+
+    arma::colvec yimp = neibo(y_full,ypred_mis,k);
+
+    return yimp;
+}
+
+static R_CallMethodDef callMethods[]  = {
+  {"neibo", (DL_FUNC) &neibo, 3},
+  {"neibo_index", (DL_FUNC) &neibo_index, 3},
+  {"pmm_neibo", (DL_FUNC) &pmm_neibo, 4},
+  {"pmm_weighted_neibo", (DL_FUNC) &pmm_weighted_neibo, 5},
+  {NULL, NULL, 0}
+};